% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moggy.R
\name{as_code}
\alias{as_code}
\alias{as_bold}
\title{Add markdown wrappers around text}
\usage{
as_code(x)

as_bold(x)
}
\arguments{
\item{x}{text}
}
\value{
text wrapped in markup
}
\description{
Add markdown wrappers around text
}
