% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moggy.R
\name{MarkdownDoc}
\alias{MarkdownDoc}
\title{Build Markdown documents programatically}
\description{
Build Markdown documents programatically

Build Markdown documents programatically
}
\examples{
\dontrun{
doc <- MarkdownDoc$new()
doc$h1("Main title")
doc$add("Lorum ipsum")
doc$table(head(mtcars))
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{doc}}{list of characters strings which make up the document}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MarkdownDoc-new}{\code{MarkdownDoc$new()}}
\item \href{#method-MarkdownDoc-add}{\code{MarkdownDoc$add()}}
\item \href{#method-MarkdownDoc-hx}{\code{MarkdownDoc$hx()}}
\item \href{#method-MarkdownDoc-h1}{\code{MarkdownDoc$h1()}}
\item \href{#method-MarkdownDoc-h2}{\code{MarkdownDoc$h2()}}
\item \href{#method-MarkdownDoc-h3}{\code{MarkdownDoc$h3()}}
\item \href{#method-MarkdownDoc-h4}{\code{MarkdownDoc$h4()}}
\item \href{#method-MarkdownDoc-table}{\code{MarkdownDoc$table()}}
\item \href{#method-MarkdownDoc-itemize}{\code{MarkdownDoc$itemize()}}
\item \href{#method-MarkdownDoc-enumerate}{\code{MarkdownDoc$enumerate()}}
\item \href{#method-MarkdownDoc-code}{\code{MarkdownDoc$code()}}
\item \href{#method-MarkdownDoc-comment}{\code{MarkdownDoc$comment()}}
\item \href{#method-MarkdownDoc-as_character}{\code{MarkdownDoc$as_character()}}
\item \href{#method-MarkdownDoc-print}{\code{MarkdownDoc$print()}}
\item \href{#method-MarkdownDoc-clone}{\code{MarkdownDoc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-new"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-new}{}}}
\subsection{Method \code{new()}}{
Create a new document
Generic way to add text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-add"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$add(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{text
Insert header}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-hx"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-hx}{}}}
\subsection{Method \code{hx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$hx(prefix, title, ref = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{Header prefix. E.g. "#"}

\item{\code{title}}{title of section}

\item{\code{ref}}{name of reference. default: NULL
Insert header}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-h1"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-h1}{}}}
\subsection{Method \code{h1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$h1(title, ref = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title of section}

\item{\code{ref}}{name of reference. default: NULL
Insert header}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-h2"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-h2}{}}}
\subsection{Method \code{h2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$h2(title, ref = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title of section}

\item{\code{ref}}{name of reference. default: NULL
Insert header}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-h3"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-h3}{}}}
\subsection{Method \code{h3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$h3(title, ref = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title of section}

\item{\code{ref}}{name of reference. default: NULL
Insert header}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-h4"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-h4}{}}}
\subsection{Method \code{h4()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$h4(title, ref = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title of section}

\item{\code{ref}}{name of reference. default: NULL
Insert a data.frame as a markdown table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-table"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-table}{}}}
\subsection{Method \code{table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$table(df, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{data.frame}

\item{\code{...}}{arguments passed to \code{knitr::kable()}
Create a itemized list from a list of items}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-itemize"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-itemize}{}}}
\subsection{Method \code{itemize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$itemize(items)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{items}}{character vector of list items
Create a numbered list from a list of items}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-enumerate"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-enumerate}{}}}
\subsection{Method \code{enumerate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$enumerate(items)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{items}}{character vector of list items
Insert an R code block}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-code"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-code}{}}}
\subsection{Method \code{code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$code(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{text to insert}

\item{\code{...}}{chunk options e.g. \code{val=FALSE, echo = TRUE}
Insert an HTML comment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-comment"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-comment}{}}}
\subsection{Method \code{comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$comment(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{comment text}

\item{\code{...}}{ignored
Convert class to character string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-as_character"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-as_character}{}}}
\subsection{Method \code{as_character()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$as_character(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored
Default print method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-print"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkdownDoc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MarkdownDoc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkdownDoc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
