% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moggy.R
\name{sec_cite}
\alias{sec_cite}
\title{Cite a reference to a particular section which}
\usage{
sec_cite(ref, label)
}
\arguments{
\item{ref}{name of reference to link to}

\item{label}{label to show in link}
}
\description{
Cite a reference to a particular section which
}
