% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moggy.R
\name{sec_ref}
\alias{sec_ref}
\title{Create a setion reference}
\usage{
sec_ref(ref)
}
\arguments{
\item{ref}{name of reference which can be linked to with \code{seg_cite()}}
}
\description{
Create a setion reference
}
